BaseCache class added
