+++ src/types/cache/base-cache.ts
@@ -1,0 +1,6 @@
+import { SetupCache, SetupCacheEntry } from "@src/types/cache/circuits-setup-cache";
+import { CompileCache, CompileCacheEntry } from "@src/types/cache/circuits-compile-cache";
+
+export type BaseCacheType = SetupCache | CompileCache;
+
+export type BaseCacheEntry = SetupCacheEntry | CompileCacheEntry;
